Moving towards being instance based, for now just a change to make DKManager a proper singleton Removed SecureUUID as it's no longer needed

+++ DeploydKit-Private/NSURLConnection+Timeout.m
@@ -102,7 +102,7 @@
   
   // Set error
   if (internalErr != nil) {
-    NSLog(@"error: %@ (%i)", internalErr.localizedDescription, internalErr.code);
+    NSLog(@"error: %@ (%li)", internalErr.localizedDescription, (long)internalErr.code);
     if (error != NULL) {
       *error = internalErr;
     }
